Added ReceiverState, MessageEvent, ExceptionEvent. Moved IService, IServiceBuilder and ControlState to specified directories, removed async boolean oriented logic from WebServer and WebServerServiceBuilder.
