+++ NetworkingUtilities/Http/Services/WebServerServiceBuilder.cs
@@ -1,4 +1,5 @@
-﻿using NetworkingUtilities.Http.Routing;
+﻿using NetworkingUtilities.Abstracts;
+using NetworkingUtilities.Http.Routing;
 
 namespace NetworkingUtilities.Http.Services
 {
@@ -27,15 +28,8 @@ namespace NetworkingUtilities.Http.Services
 			return this;
 		}
 
-		public WebServerServiceBuilder UseAsyncInvocations(bool value)
-		{
-			Async = value;
-			return this;
-		}
-
 		internal IRouter Router { get; }
 		internal int? Port { get; private set; }
 		internal string Prefix { get; private set; }
-		internal bool Async { get; private set; }
 	}
 }
