+++ NetworkingUtilities/Utilities/Events/MessageEvent.cs
@@ -1,0 +1,18 @@
+﻿using System;
+
+namespace NetworkingUtilities.Utilities.Events
+{
+	public class MessageEvent : EventArgs
+	{
+		public string Message { get; }
+		public string From { get; }
+		public string To { get; }
+
+		public MessageEvent(string message, string from, string to)
+		{
+			Message = message;
+			From = from;
+			To = to;
+		}
+	}
+}
